+++ detectors/autodetect/autodetect.go
@@ -21,6 +21,7 @@ import (
 	"go.opentelemetry.io/contrib/detectors/aws/lambda"
 	"go.opentelemetry.io/contrib/detectors/azure/azurecontainerapps"
 	"go.opentelemetry.io/contrib/detectors/azure/azurevm"
+	"go.opentelemetry.io/contrib/detectors/docker"
 	"go.opentelemetry.io/contrib/detectors/gcp"
 	"go.opentelemetry.io/contrib/detectors/hetzner"
 	"go.opentelemetry.io/contrib/detectors/k8sapi"
@@ -133,6 +134,10 @@ var (
 	// identifying the container in which the process is running, especially in
 	// containerized environments like Kubernetes or Docker.
 	IDContainer = ID("container")
+	// IDDocker is the ID for the Docker detector that detects resource
+	// attributes on Docker containers (see docker.NewResourceDetector for
+	// details).
+	IDDocker = ID("docker")
 )
 
 var (
@@ -174,6 +179,8 @@ var (
 		IDProcessRuntimeDescription: optFactory(resource.WithProcessRuntimeDescription()),
 
 		IDContainer: optFactory(resource.WithContainer()),
+
+		IDDocker: func() resource.Detector { return docker.NewResourceDetector() },
 	}
 )